requirements.txt
sentence_transformers
faiss-cpu
python-telegram-bot
python-dotenv
elevenlabs
openai
arabic-reshaper
python-bidi
watchdog
camel-tools